+++ CSGOInternal/SetYawOffsetHook.cpp
@@ -30,7 +30,7 @@ SetYawOffsetHook::SetYawOffsetHook(uintptr_t moduleBaseIn)
 
 bool SetYawOffsetHook::initialize()
 {
-	printf("Hooking %s function at %#8x\n", hookName.c_str(), SetYawOffsetHook::oFunction);
+	printf("Hooking %s function at %#8x\n", hookName.c_str(), (uintptr_t)SetYawOffsetHook::oFunction);
 	SetYawOffsetHook::oFunction = (tTargetPtr)trampoline32((uintptr_t)SetYawOffsetHook::oFunction, (uintptr_t)SetYawOffsetHook::hookFunction, hookLen);
 	enabled = true;
 	return true;
